+++ java/src/hype/core/util/HMath.java
@@ -524,6 +524,41 @@ public class HMath implements HConstants {
 	}
 	
 	
+	// COMPARATORS W/ TOLERANCE //
+	
+	public static boolean lessThan(float a, float b, float tolerance) {
+		return a < b + tolerance;
+	}
+	
+	public static boolean lessThan(float a, float b) {
+		return a < b + TOLERANCE;
+	}
+	
+	public static boolean greaterThan(float a, float b, float tolerance) {
+		return b < a + tolerance;
+	}
+	
+	public static boolean greaterThan(float a, float b) {
+		return b < a + TOLERANCE;
+	}
+	
+	public static boolean isEqual(float a, float b, float tolerance) {
+		return Math.abs(a-b) < tolerance;
+	}
+	
+	public static boolean isEqual(float a, float b) {
+		return Math.abs(a-b) < TOLERANCE;
+	}
+	
+	public static boolean isZero(float a, float tolerance) {
+		return Math.abs(a) < tolerance;
+	}
+	
+	public static boolean isZero(float a) {
+		return Math.abs(a) < TOLERANCE;
+	}
+	
+	
 	// MISC //
 	
 	public static float map(float val,